init TypeSpecificCallable, TypeProjection
