+++ src/pformat/__init__.py
@@ -1,9 +1,3 @@
-from .common_types import (
-    MultilineTypeFormatterFunc,
-    NormalTypeFormatterFunc,
-    TypeProjectionFunc,
-    TypeProjectionFuncMapping,
-)
 from .format_options import FormatOptions
 from .indentation_utility import IndentMarker, IndentType
 from .pretty_formatter import DefaultFormatter, IterableFormatter, MappingFormatter, PrettyFormatter
@@ -14,12 +8,15 @@ from .text_style import (
     rm_style_modifiers,
     strlen_no_style,
 )
-from .type_specific_formatters import (
+from .type_formatters import (
     CustomMultilineFormatter,
     CustomNormalFormatter,
     MultilineFormatter,
+    MultilineTypeFormatterFunc,
     NormalFormatter,
+    NormalTypeFormatterFunc,
     TypeFormatter,
     multiline_formatter,
     normal_formatter,
 )
+from .type_projections import TypeProjectionFunc, TypeProjectionFuncMapping
