+++ test/test_type_formatters.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import pytest
 
-from pformat.type_specific_formatters import (
+from pformat.type_formatters import (
     CustomMultilineFormatter,
     CustomNormalFormatter,
     MultilineFormatter,
@@ -72,7 +72,7 @@ class TestTypeFormatterCommon:
     def test_repr(self, sut):
         assert repr(sut) == f"{self.fmt_type.__name__}({self.type.__name__})"
 
-    @patch("pformat.type_specific_formatters.has_valid_type")
+    @patch("pformat.type_specific_callable.has_valid_type")
     def test_has_valid_type(self, mock_has_valid_type, sut: TypeFormatter):
         class DummyType:
             pass
